+++ src/main/java/swp391_gr7/hivsystem/model/Reports.java
@@ -23,7 +23,7 @@ public class Reports {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "report_id")
-    private Long reportID;
+    private int reportID;
 //
     @ManyToOne(fetch = FetchType.LAZY)
